Added robots.txt checker functionality
A new file, robots_checker.py, was created to implement a utility for checking the robots.txt file of a given domain. This utility fetches the robots.txt file and determines the rules applicable to our bot, "GPTBot". Furthermore, the related route was added to the views.py file and the appropriate response schema, "CheckRobotsResult", was added to schemas.py. These updates facilitate a new API endpoint which allows users to check if GPTBot is allowed to crawl a specific domain. The functionality was added to provide more transparency for users and avoid unnecessary HTTP error codes due to crawling disallowed domains.