Add robots url check and update dependencies
Added a method to check robots permissions by URL in the api/views.py. An additional import is added for check_robots_url. Also, to further solve the problem, additional dependencies such as "certifi", "charset-normalizer" and "requests" packages are added to poetry.lock file. This update will ensure that the program will correctly handle URLs and its permissions.

A new test file is added to aid in testing the robots URL check. This would help in maintaining the code quality and make sure everything works as expected.