Add Gunicorn support and change default redirect for project
Added Gunicorn to the project and its dependencies as a WSGI HTTP Server for Unix operating systems. This inclusion allows for better handling of multiple simultaneous requests and generally improves the performance of the server.
An additional trusted domain was added to the settings for wider accessibility of the application.
In addition, changed the default redirect from "dashboard/" to "api/" in urls.py to improve user navigation. This enhances user experience as the API is often the main entry point for users interacting programmatically with our service.